Remove site-root parameter. This coupling isn't required.
Removing this gives us the ability to use blowgun in environments that
don't specifically have a website running in them. For example, you
might just want to run a remote command on a server.

Current scripts don't actually use the "webroot" parameter anyway, so we
won't really miss this site-root setting.