Extend the visibility period to last until the next timeout.
Previously, with visibility period = timeout, there was a small chance
for the node to not manage to update the visibility in time, and the
essage to be picked by other nodes in that split second.

Also, extend the visibility period in general to 120 seconds, timeout to
half of that and log a notice whenever the visibility is extended.